[🛠️Fix] Stat Data
- 스탯 자료형 데이터 수정 int => ulong

+++ IdleGame/Assets/Scripts/Main/Model/Enemy/BaseEnemy.cs
@@ -11,18 +11,18 @@ public class BaseEnemy : MonoBehaviour, IDamageable
     private Coroutine _attackCoroutine;
 
     //체력
-    private int _maxHp;
-    private int _currentHP;
+    private ulong _maxHp;
+    private ulong _currentHP;
 
     //공격
-    private int _damage;
+    private ulong _damage;
     private float _attackSpeed;
     private float _range;
 
     //속도
     private float _moveSpeed;
 
-    public int _rewards;
+    public ulong _rewards;
 
     private SpriteRenderer _spriteRenderer;
     private Rigidbody2D _rigidbody;
@@ -56,7 +56,7 @@ public class BaseEnemy : MonoBehaviour, IDamageable
 
     public void SetStatWeight(int Weight)
     {
-        int _weight = Weight - 1;
+        ulong _weight = (ulong)(Weight - 1);
         _maxHp = _maxHp + _maxHp * _weight;
         _damage = _damage + _damage * _weight;
         _rewards = _rewards + _rewards * _weight;
@@ -134,12 +134,12 @@ public class BaseEnemy : MonoBehaviour, IDamageable
         }
     }
 
-    public void TakeDamage(int Damage)
+    public void TakeDamage(ulong Damage)
     {
         AmountDamage(Damage);
         Debug.Log($"{gameObject.name} : {_currentHP}");
     }
-    private void AmountDamage(int Damage)
+    private void AmountDamage(ulong Damage)
     {
         if (_currentHP - Damage <= 0)
         {
