+++ IdleGame/Assets/Scripts/Main/Model/InterfaceEntity/IDamageable.cs
@@ -5,5 +5,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeDamage(int Damage);
+    void TakeDamage(ulong Damage);
 }
